requirements.txt
discord.py==1.5.1
python-dotenv==0.14.0
pynacl==1.4.0
requests==2.24.0
youtube_dl==2020.9.20
requests_html==0.10.0
beautifulsoup4==4.9.3